requirements.txt
pandas
matplotlib
python-docx
psycopg2-binary
